backend/requirements.txt
appdirs==1.4.4
asgiref==3.9.1
attrs==25.3.0
Babel==2.10.3
beautifulsoup4==4.12.3
cached-property==2.0.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
channels==4.2.2
channels_redis==4.3.0 
msgpack==1.1.1 
redis==6.2.0
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==42.0.0
daphne==4.2.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.2
django-bootstrap4==24.4
django-ckeditor==6.7.2
django-classy-tags==4.1.0
django-cleanup==9.0.0
django-cms==4.1.4
django-cors-headers==4.7.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==25.1
django-formtools==2.5.1
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-mathfilters==1.0.0
django-sekizai==4.1.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangocms-admin-style==3.3.1
djangocms-text-ckeditor==5.1.6
djangorestframework==3.16.0
djangorestframework-simplejwt>=5.3.0
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary>=2.9.9
Pillow>=10.3.0
docopt==0.6.2
docutils==0.20.1
drf-spectacular==0.28.0
ebaysdk==2.1.5
filetype==1.2.0
freezegun==1.2.1
geoip2==2.9.0
gevent==25.5.1
greenlet==3.2.3
html5lib==1.1
idna==3.6
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
libsass==0.22.0
lxml==4.9.3
MarkupSafe==2.1.2
maxminddb==2.7.0
num2words==0.5.10
ofxparse==0.21
openai
packaging==24.2
passlib==1.7.4
Pillow>=10.0.0
platformdirs==4.3.8
polib==1.1.1
psutil==5.9.4
psycopg2-binary>=2.9.9
pycparser==2.22
pydot==1.4.2
pyOpenSSL==21.0.0
pyparsing==3.2.3
PyPDF2==2.12.1
pypng==0.20220715.0
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.8
python-stdnum==1.19
pytz==2025.2
pyusb==1.2.1
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
reportlab>=4.0.0
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rjsmin
rpds-py
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.5
vobject==0.9.6.1
webencodings==0.5.1
Werkzeug==2.2.2
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.2.1
zope.event==5.0
zope.interface==7.2
openpyxl>=3.1.0
pandas>=2.0.0
